clean up and refactor send-bitcoin-details-screen
