+++ app/components/send-flow/DetailAmountNote.tsx
@@ -1,0 +1,181 @@
+import React from "react"
+import { View } from "react-native"
+import { makeStyles, Text } from "@rneui/themed"
+import { useI18nContext } from "@app/i18n/i18n-react"
+import { useAppConfig, useBreez } from "@app/hooks"
+
+// components
+import { GaloyTertiaryButton } from "@app/components/atomic/galoy-tertiary-button"
+import { AmountInput } from "@app/components/amount-input/amount-input"
+import { NoteInput } from "@app/components/note-input"
+
+// types
+import { PaymentDetail } from "@app/screens/send-bitcoin-screen/payment-details"
+import { WalletCurrency } from "@app/graphql/generated"
+import {
+  MoneyAmount,
+  toBtcMoneyAmount,
+  toUsdMoneyAmount,
+  WalletOrDisplayCurrency,
+} from "@app/types/amounts"
+
+// utils
+import { testProps } from "../../utils/testProps"
+
+type Props = {
+  usdWallet: any
+
+  paymentDetail: PaymentDetail<WalletCurrency>
+  setPaymentDetail: (val: PaymentDetail<WalletCurrency>) => void
+}
+
+const DetailAmountNote: React.FC<Props> = ({
+  usdWallet,
+  paymentDetail,
+  setPaymentDetail,
+}) => {
+  const styles = useStyles()
+  const { LL } = useI18nContext()
+  const { btcWallet } = useBreez()
+  const { appConfig } = useAppConfig()
+
+  const { sendingWalletDescriptor } = paymentDetail
+
+  const lnurlParams =
+    paymentDetail?.paymentType === "lnurl" ? paymentDetail?.lnurlParams : undefined
+
+  const sendAll = () => {
+    let moneyAmount: MoneyAmount<WalletCurrency>
+
+    if (paymentDetail.sendingWalletDescriptor.currency === WalletCurrency.Btc) {
+      moneyAmount = {
+        amount: btcWallet.balance ?? 0,
+        currency: WalletCurrency.Btc,
+        currencyCode: "BTC",
+      }
+    } else {
+      moneyAmount = {
+        amount: usdWallet?.balance ?? 0,
+        currency: WalletCurrency.Usd,
+        currencyCode: "USD",
+      }
+    }
+
+    setPaymentDetail(
+      paymentDetail?.setAmount
+        ? paymentDetail.setAmount(moneyAmount, true)
+        : paymentDetail,
+    )
+  }
+
+  const setAmount = (moneyAmount: MoneyAmount<WalletOrDisplayCurrency>) => {
+    setPaymentDetail(
+      paymentDetail?.setAmount ? paymentDetail.setAmount(moneyAmount) : paymentDetail,
+    )
+  }
+
+  return (
+    <>
+      <View style={styles.fieldContainer}>
+        <View style={styles.amountRightMaxField}>
+          <Text {...testProps(LL.SendBitcoinScreen.amount())} style={styles.amountText}>
+            {LL.SendBitcoinScreen.amount()}
+          </Text>
+          {paymentDetail.canSendMax && !paymentDetail.isSendingMax && (
+            <GaloyTertiaryButton
+              clear
+              title={LL.SendBitcoinScreen.maxAmount()}
+              onPress={sendAll}
+            />
+          )}
+        </View>
+        <View style={styles.currencyInputContainer}>
+          <AmountInput
+            unitOfAccountAmount={paymentDetail.unitOfAccountAmount}
+            setAmount={setAmount}
+            convertMoneyAmount={paymentDetail.convertMoneyAmount}
+            walletCurrency={sendingWalletDescriptor.currency}
+            canSetAmount={paymentDetail.canSetAmount}
+            isSendingMax={paymentDetail.isSendingMax}
+            maxAmount={
+              appConfig.galoyInstance.name === "Staging"
+                ? {
+                    amount: 2500,
+                    currency: "DisplayCurrency",
+                    currencyCode: "USD",
+                  }
+                : lnurlParams?.max
+                ? toBtcMoneyAmount(lnurlParams.max)
+                : undefined
+            }
+            minAmount={
+              lnurlParams?.min
+                ? toUsdMoneyAmount(lnurlParams.min)
+                : {
+                    amount: 1,
+                    currency: "USD",
+                    currencyCode: "USD",
+                  }
+            }
+          />
+        </View>
+      </View>
+      {paymentDetail.paymentType === "intraledger" && (
+        <View style={styles.fieldContainer}>
+          <Text style={styles.fieldTitleText}>{LL.SendBitcoinScreen.note()}</Text>
+          <NoteInput
+            onChangeText={(text) =>
+              paymentDetail.setMemo && setPaymentDetail(paymentDetail.setMemo(text))
+            }
+            value={paymentDetail.memo || ""}
+            editable={paymentDetail.canSetMemo}
+          />
+        </View>
+      )}
+    </>
+  )
+}
+
+export default DetailAmountNote
+
+const useStyles = makeStyles(({ colors }) => ({
+  sendBitcoinAmountContainer: {
+    flex: 1,
+  },
+
+  fieldTitleText: {
+    fontWeight: "bold",
+    marginBottom: 4,
+  },
+  fieldContainer: {
+    marginBottom: 12,
+  },
+  currencyInputContainer: {
+    flexDirection: "column",
+  },
+
+  buttonContainer: {
+    flex: 1,
+    justifyContent: "flex-end",
+  },
+  modal: {
+    marginBottom: "90%",
+  },
+  pickWalletIcon: {
+    marginRight: 12,
+  },
+  screenStyle: {
+    padding: 20,
+    flexGrow: 1,
+  },
+  amountText: {
+    fontWeight: "bold",
+  },
+  amountRightMaxField: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+    marginBottom: 8,
+    height: 18,
+  },
+}))
