+++ app/components/send-flow/index.ts
@@ -1,3 +1,6 @@
 import DestinationField from "./DestinationField"
+import ChooseWallet from "./ChooseWallet"
+import DetailDestination from "./DetailDestination"
+import DetailAmountNote from "./DetailAmountNote"
 
-export { DestinationField }
+export { DestinationField, ChooseWallet, DetailDestination, DetailAmountNote }
